делаю модульность
