+++ js/util.js
@@ -1,0 +1,12 @@
+function checkLength(string, maxLengh){
+  return string.length <= maxLengh;
+}
+checkLength('asd',4);
+
+function getRandomInt(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min)) + min; //Максимум не включается, минимум включается
+}
+
+export {getRandomInt};
